+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/FlinkDeploymentManagerProvider.scala
@@ -152,7 +152,7 @@ object FlinkStreamingPropertiesConfig {
     ScenarioPropertyConfig(
       defaultValue = None,
       editor = Some(StaticStringParameterEditor),
-      validators = Some(List(LiteralIntegerValidator, MinimalNumberValidator(1))),
+      validators = Some(List(LiteralIntegerExpressionValidator, MinimalNumberValidator(1))),
       label = Some("Parallelism"),
       hintText = None
     )
@@ -173,7 +173,7 @@ object FlinkStreamingPropertiesConfig {
     ScenarioPropertyConfig(
       defaultValue = None,
       editor = Some(FixedValuesParameterEditor(spillStatePossibleValues)),
-      validators = Some(List(FixedValuesValidator(spillStatePossibleValues))),
+      validators = Some(List(FixedValuesExpressionValidator(spillStatePossibleValues))),
       label = Some("Spill state to disk"),
       hintText = None
     )
@@ -182,7 +182,7 @@ object FlinkStreamingPropertiesConfig {
     ScenarioPropertyConfig(
       defaultValue = None,
       editor = Some(FixedValuesParameterEditor(asyncPossibleValues)),
-      validators = Some(List(FixedValuesValidator(asyncPossibleValues))),
+      validators = Some(List(FixedValuesExpressionValidator(asyncPossibleValues))),
       label = Some("IO mode"),
       hintText = None
     )
@@ -191,7 +191,7 @@ object FlinkStreamingPropertiesConfig {
     ScenarioPropertyConfig(
       defaultValue = None,
       editor = Some(StaticStringParameterEditor),
-      validators = Some(List(LiteralIntegerValidator, MinimalNumberValidator(1))),
+      validators = Some(List(LiteralIntegerExpressionValidator, MinimalNumberValidator(1))),
       label = Some("Checkpoint interval in seconds"),
       hintText = None
     )